requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pillow==11.0.0
pytesseract==0.3.13
python-multipart==0.0.20
